+++ modules/features/endofyear/src/main/java/au/com/shiftyjelly/pocketcasts/endofyear/ui/StoriesPage.kt
@@ -274,7 +274,6 @@ internal fun BoxScope.TopControls(
             state = pagerState,
             progress = progress,
             activeColor = Color.Black,
-            isProgressedActive = false,
         )
         Spacer(
             modifier = Modifier.height(10.dp),