+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 *   Bug Fixes
     *   Fix effects being lost when paused
         ([#3219](https://github.com/Automattic/pocket-casts-android/pull/3219))
+*   Updates
+    *   Update profile header UI
+        ([#3235](https://github.com/Automattic/pocket-casts-android/pull/3235))
 
 7.77
 -----
@@ -30,6 +33,9 @@
 *   Bug Fixes
     *   Use red color for the notification icons
         ([#3154](https://github.com/Automattic/pocket-casts-android/pull/3154))
+    *   Fix effects being lost when paused
+        ([#3219](https://github.com/Automattic/pocket-casts-android/pull/3219))
+
 
 7.76
 -----